Переносит функции отрисовки и создания элемента в отдельный файл
